Update drf validation tests
