+++ tests/rest_framework/test_filterset.py
@@ -4,7 +4,9 @@ from django.conf import settings
 from django.test import TestCase
 from django.test.utils import override_settings
 
+from django_filters import STRICTNESS
 from django_filters.compat import is_crispy
+from rest_framework.exceptions import ValidationError
 from django_filters.rest_framework import FilterSet, filters
 from django_filters.widgets import BooleanWidget
 
@@ -43,3 +45,25 @@ class CrispyFormsCompatTests(TestCase):
     def test_form_initialization(self):
         # ensure that crispy compat does not prematurely initialize the form
         self.assertFalse(hasattr(ArticleFilter(), '_form'))
+
+
+class ValidationErrorTests(TestCase):
+
+    def test_simple(self):
+        class F(FilterSet):
+            class Meta:
+                model = Article
+                fields = ['id', 'author', 'name']
+                strict = STRICTNESS.RAISE_VALIDATION_ERROR
+
+        f = F(data={'id': 'foo', 'author': 'bar', 'name': 'baz'})
+        with self.assertRaises(ValidationError) as exc:
+            f.qs
+
+        # test rendered output, includes error code
+        self.assertDictEqual(exc.exception.get_full_details(), {
+            'id': [{'message': 'Enter a number.', 'code': 'invalid'}],
+            'author': [
+                {'message': 'Select a valid choice. That choice is not one of the available choices.', 'code': 'invalid_choice'},
+            ],
+        })
