+++ tests/test_utils.py
@@ -5,7 +5,7 @@ import django
 from django.db import models
 from django.db.models.constants import LOOKUP_SEP
 from django.db.models.fields.related import ForeignObjectRel
-from django.forms import ValidationError
+from django.forms import ValidationError as DjValidationError
 from django.test import TestCase, override_settings
 from django.utils.functional import Promise
 from django.utils.timezone import get_default_timezone
@@ -17,8 +17,8 @@ from django_filters.utils import (
     get_model_field,
     handle_timezone,
     label_for_filter,
-    raw_validation,
     resolve_field,
+    translate_validation,
     verbose_field_name,
     verbose_lookup_expr
 )
@@ -350,19 +350,17 @@ class HandleTimezone(TestCase):
         self.assertEqual(handled, get_default_timezone().localize(dst_starting_date, True))
 
 
-class RawValidationDataTests(TestCase):
+class TranslateValidationDataTests(TestCase):
+
     def test_simple(self):
-        class F(FilterSet):
-            class Meta:
-                model = Article
-                fields = ['id', 'author', 'name']
-                strict = STRICTNESS.RAISE_VALIDATION_ERROR
+        error = DjValidationError({
+            'id': [DjValidationError('Enter a number.', 'invalid')],
+            'author': [DjValidationError('Select a valid choice. That choice is not one of the available choices.', 'invalid_choice')],
+        })
 
-        f = F(data={'id': 'foo', 'author': 'bar', 'name': 'baz'})
-        with self.assertRaises(ValidationError) as exc:
-            f.qs
-
-        self.assertDictEqual(raw_validation(exc.exception), {
-            'id': ['Enter a number.'],
-            'author': ['Select a valid choice. That choice is not one of the available choices.'],
+        self.assertDictEqual(translate_validation(error).get_full_details(), {
+            'id': [{'message': 'Enter a number.', 'code': 'invalid'}],
+            'author': [
+                {'message': 'Select a valid choice. That choice is not one of the available choices.', 'code': 'invalid_choice'},
+            ],
         })
